refactor: 클래스 위치 변경 domain -> domain.type
Enum 클래스는 별도의 type이라는 패키지 내에서 관리하기로 하였다.
